+++ App/Models/Spells/LesserHealingWave.cs
@@ -88,7 +88,7 @@ namespace App.Models.Spells
 
             foreach (var modifier in healingModifiers)
             {
-                aa = (int)aa * modifier.Value;
+                aa = aa * modifier.Value;
             }
 
             return (int?)aa;
@@ -102,7 +102,7 @@ namespace App.Models.Spells
 
             foreach (var modifier in healingModifiers)
             {
-                aa = (int)aa * modifier.Value;
+                aa = aa * modifier.Value;
             }
 
             return (int?)aa;
