+++ App/Models/Spells/Riptide.cs
@@ -105,7 +105,7 @@ namespace App.Models.Spells
 
             foreach (var modifier in healingModifiers)
             {
-                aa = (int)aa * modifier.Value;
+                aa = aa * modifier.Value;
             }
 
             return (int?)aa;
@@ -119,7 +119,7 @@ namespace App.Models.Spells
 
             foreach (var modifier in healingModifiers)
             {
-                aa = (int)aa * modifier.Value;
+                aa = aa * modifier.Value;
             }
 
             return (int?)aa;
